Make parsing work for the nested records; We now have adress and person information.
This is not the fanciest of commit's but there is an apperant problem in
parsing and looping over XML snippets. When looping over an xpath extracted
part of the XML it is imporntant that the snippet has namespace definitions for
all namespaces used within that block, otherwise that part of the XML cannot be
processed further with the xpath() calls within plpgsql. This will be the case
for the adress blocks here, if looping over the logical parent we cannot
process information within the children. Hence the extremely fancy workaround
using a numbered subindex.